agents/backend_ads/agents/backend/onyx/server/features/modules/optimization/requirements.txt
# Optimization Module Requirements
# Core dependencies for high-performance system optimization

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data Validation and Settings
pydantic>=2.4.0
python-decouple>=3.8

# Async and Utilities
asyncio>=3.11.0
structlog>=23.2.0

# High-Performance Serialization
orjson>=3.9.0  # Rust-based JSON (10x faster)
msgpack>=1.0.0  # Binary serialization
ujson>=5.8.0   # Fallback fast JSON

# High-Performance Hashing
xxhash>=3.4.0   # Extremely fast hashing
blake3>=0.3.0   # Cryptographic hashing

# Compression
lz4>=4.3.0      # Ultra-fast compression
blosc2>=2.0.0   # Scientific data compression
zstandard>=0.22.0  # Facebook's compression

# Database Optimization
asyncpg>=0.29.0     # Fast PostgreSQL driver
aiopg>=1.4.0        # Alternative PostgreSQL driver
aiomysql>=0.2.0     # MySQL async driver
asyncio-redis>=0.16.0  # Redis async client
aioredis>=2.0.0     # Modern Redis client

# Network Optimization
aiohttp>=3.9.0      # Async HTTP client/server
httpx>=0.25.0       # Modern HTTP client
h2>=4.1.0           # HTTP/2 support
aiodns>=3.1.0       # Fast DNS resolution

# Memory and System Optimization
psutil>=5.9.0       # System monitoring
pympler>=0.9.0      # Memory profiling
memory-profiler>=0.61.0  # Memory usage tracking

# Numerical and Math Optimization
numpy>=1.24.0       # Numerical computing
numba>=0.58.0       # JIT compilation
numexpr>=2.8.0      # Fast numerical expressions
bottleneck>=1.3.0   # Fast numpy operations

# GPU Acceleration (Optional)
# cupy-cuda12x>=12.0.0    # CUDA support
# torch>=2.1.0            # PyTorch GPU

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Monitoring and Metrics
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Optional Advanced Features
# Uncomment if needed for specific optimizations

# Advanced Math Libraries
# scipy>=1.11.0
# scikit-learn>=1.3.0

# Advanced Compression
# python-snappy>=0.6.0
# python-lzo>=1.14

# Advanced Hashing
# pyhash>=0.9.0
# hashlib-compat>=1.0.0

# System-Level Optimization
# setproctitle>=1.3.0
# resource>=0.2.1

# Message Queues
# aiokafka>=0.9.0
# aio-pika>=9.3.0

# Advanced Async
# trio>=0.23.0
# anyio>=4.1.0

# Database Extensions
# sqlalchemy>=2.0.0
# alembic>=1.12.0

# Caching Extensions
# cashews>=6.1.0
# aiocache>=0.12.0

# Security
# cryptography>=41.0.0

# Additional Utilities
# python-dateutil>=2.8.0
# pytz>=2023.3
# pyyaml>=6.0.0 